+++ examples/docs/changelog.md
@@ -9,8 +9,10 @@
 #### Features
 
 - input: 支持v-model双向绑定
-- 增加 `form-tip` `form-opr` `from-extra` 组件
+- 增加 `form-tip` `form-opr` `from-extra` `checkbox-agree` 组件
 - input: 增加error错误状态
+- button: 增加inForm属性
+- icon: 增加mini属性
 
 ### 0.1.0-beta